+++ ingest/operation_details.go
@@ -388,6 +388,18 @@ func (is *Session) operationDetails() map[string]interface{} {
 	case xdr.OperationTypeCloseSwap:
 	case xdr.OperationTypeCreateData:
 	case xdr.OperationTypeUpdateData:
+	case xdr.OperationTypeUpdateDataOwner:
+		op := c.Operation().Body.MustUpdateDataOwnerOp()
+		details["new_owner"] = op.NewOwner
+		details["source"] = c.Operation().SourceAccount
+	case xdr.OperationTypeCreateDataOwnerUpdateRequest:
+		op := c.Operation().Body.MustCreateDataOwnerUpdateRequestOp()
+		details["data_id"] = op.DataOwnerUpdateRequest.UpdateDataOwnerOp.DataId
+		details["new_owner"] = op.DataOwnerUpdateRequest.UpdateDataOwnerOp.NewOwner
+	case xdr.OperationTypeCancelDataOwnerUpdateRequest:
+		op := c.Operation().Body.MustCancelDataOwnerUpdateRequestOp()
+		details["request_id"] = op.RequestId
+
 	case xdr.OperationTypeRemoveData:
 	case xdr.OperationTypeCreateRedemptionRequest:
 		op := c.Operation().Body.MustCreateRedemptionRequestOp()